feat: allow selected rows to be exported
